Edit the side bar avatar to show a thinking common man
The idea is that the blog truly expects the reader to be a common man but also
the man is a thinker. We may not be equipped with the insider information and
lofty contacts but we are equipped well enought to think independently and
come up with ways that do well for ourselves in the long run.