Preface for Adam's Interview.
This is a new series that I'm testing so it can serve people's
requests. I received a lot of requests that mentioned that they were
actually happy with the content of the podcast itself but simply
lacked the attention span to keep up with the content, especially when
the guest and I go into the weeds. Also, I'd like to keep updating
regularly so that the audience is engaged on a regular basis.

So I'm testing this new format where the overall content is split into
smaller chunks or parts of larger interview. Each part will be
standalone in it's content, say, where the guest answers one
topic. This could be about their work setup, or about an eye-opening
experience they might have had, or a gig they did a while back.

Although, this means more work for me to edit the content more
regularly, hopefully, this will be more interesting for you, the
listener. Please let me know by leaving a comment on
www.voidstarpocast.com or iTunes or Google Podcasts or wherever else
you consume this. Thanks and enjoy my conversation with Adam!